remove warning

+++ NFComm/NFCore/NFDataList.hpp
@@ -473,7 +473,7 @@ public:
 
 	virtual const NF_SHARE_PTR<NFData> GetStack(const int index) const
 	{
-		if (index < mvList.size())
+		if (index < (int)mvList.size())
 		{
 			return mvList[index];
 		}
@@ -663,7 +663,7 @@ public:
     
 	virtual bool Add(const NFINT64 value)
 	{
-		if (GetCount() == mvList.size())
+		if (GetCount() == (int)mvList.size())
 		{
 			AddStatck();
 		}
@@ -682,7 +682,7 @@ public:
 
 	virtual bool Add(const double value)
 	{
-		if (GetCount() == mvList.size())
+		if (GetCount() == (int)mvList.size())
 		{
 			AddStatck();
 		}
@@ -701,7 +701,7 @@ public:
 
 	virtual bool Add(const std::string& value)
 	{
-		if (GetCount() == mvList.size())
+		if (GetCount() == (int)mvList.size())
 		{
 			AddStatck();
 		}
@@ -720,7 +720,7 @@ public:
 
 	virtual bool Add(const NFGUID& value)
 	{
-		if (GetCount() == mvList.size())
+		if (GetCount() == (int)mvList.size())
 		{
 			AddStatck();
 		}
@@ -739,7 +739,7 @@ public:
 
 	virtual bool Add(const NFVector2& value)
 	{
-		if (GetCount() == mvList.size())
+		if (GetCount() == (int)mvList.size())
 		{
 			AddStatck();
 		}
@@ -758,7 +758,7 @@ public:
 
 	virtual bool Add(const NFVector3& value)
 	{
-		if (GetCount() == mvList.size())
+		if (GetCount() == (int)mvList.size())
 		{
 			AddStatck();
 		}
